* fixed warning C4267: conversion from 'size_t' to '__u32', possible loss of data (71) ** add BUG_ON() if there is a possibility of overflow ** add forced Type Casting

+++ drbd/drbd_nl.c
@@ -7504,7 +7504,7 @@ void notify_helper(enum drbd_notification_type type,
 #else
 	strlcpy(helper_info.helper_name, name, sizeof(helper_info.helper_name));
 #endif
-	helper_info.helper_name_len = min(strlen(name), sizeof(helper_info.helper_name));
+	helper_info.helper_name_len = (__u32)(min(strlen(name), sizeof(helper_info.helper_name)));
 	helper_info.helper_status = status;
 
 	skb = genlmsg_new(NLMSG_GOODSIZE, GFP_NOIO);
